OnPoint-023 Multiple fixes (see details) [GM]
- Fix tests crashing due to missing foreign key pointing to the
  Moodle API. Fixed by inserting the moodle API data into the
  in-memory DB in the @Before test function.
- Use Long all over the place for the Moodle ID to match with the
  SQLite RowID
- Refresh the lists after each assignment set from a given moodle has
  been processed directly within the callback of the JSON response
  processing. This is required as the HTTP request and response are
  async and there is no way to wait for the loop of requests to
  end synchronously and update the list only once at the end.

+++ app/src/main/java/at/tugraz/onpoint/database/Database.kt
@@ -151,7 +151,7 @@ interface AssignmentDao {
         description: String,
         deadline: Long,
         links: String,
-        moodleId: Int?,
+        moodleId: Long?,
         isCustom: Boolean,
         isCompleted: Boolean,
         courseIdFromMoodle: Int?,
@@ -163,7 +163,7 @@ interface AssignmentDao {
         description: String,
         deadline: Date,
         links: List<URL>? = null,
-        moodleId: Int,
+        moodleId: Long,
         courseIdFromMoodle: Int,
         assignmentIdFromMoodle: Int,
     ): Long {
